Finalized the OrdersForProductView.xaml

+++ CarSparePartService/Interfaces/IProductFetcher.cs
@@ -3,7 +3,7 @@
 public interface IProductFetcher
 {
     IEnumerable<Product.Product> GetAllProducts();
-    Product.Product FindProduct(int productId);
+    Product.Product FindProduct(long productId);
     void LoadProducts(string fileName);
     void LoadProductsFromBackup();
 }
